Update options.html with note for macOS users
For reminders to appear for macOS users they need to allow Firefox to
send notifications, this permission appears to be disabled by default
on macOS and Firefox does not remind the user of this when the add-on
is installed.

This can cause some confusion, because Firefox will notify the user
that the add-on will have (browser-level) notification permissions on
installation, but Firefox will not notify the user if it does not have
system permissions for notifications on macOS (as is the case by
default).

I updated options.html with a note about this to prevent confusion for
macOS users. I think it might be possible to update options.html from
options.js based on OS when the page is opened so that this note is
only shown to macOS users, but this commit does not include that.